chore: inputs from code review
