+++ lib/src/bottom_sheet/theme/default_sbb_bottom_sheet_theme_data.dart
@@ -19,7 +19,7 @@ class DefaultSBBBottomSheetThemeData extends SBBBottomSheetThemeData {
           barrierColor: const Color(0x80000000),
           titleBodyGap: SBBSpacing.small,
           titleMinHeight: SBBSpacing.xLarge,
-          padding: EdgeInsets.symmetric(horizontal: SBBSpacing.medium).copyWith(top: SBBSpacing.small),
+          padding: .symmetric(horizontal: SBBSpacing.medium).copyWith(top: SBBSpacing.small),
         ),
       );
 }
