+++ lib/src/status/theme/default_sbb_status_theme_data.dart
@@ -5,28 +5,28 @@ class DefaultSBBStatusThemeData extends SBBStatusThemeData {
     : super(
         alert: SBBStatusStyle(
           textStyle: SBBTextStyles.smallLight,
-          foregroundColor: baseStyle.themeValue(SBBColors.black, SBBColors.white),
+          foregroundColor: baseStyle.colorScheme.defaultTextColor,
           backgroundColor: baseStyle.colorScheme.errorColor,
           borderColor: baseStyle.colorScheme.errorColor,
           iconColor: SBBColors.white,
         ),
         warning: SBBStatusStyle(
           textStyle: SBBTextStyles.smallLight,
-          foregroundColor: baseStyle.themeValue(SBBColors.black, SBBColors.white),
+          foregroundColor: baseStyle.colorScheme.defaultTextColor,
           backgroundColor: baseStyle.themeValue(SBBColors.peach, SBBColors.peachDark),
           borderColor: baseStyle.themeValue(SBBColors.peach, SBBColors.peachDark),
           iconColor: SBBColors.white,
         ),
         success: SBBStatusStyle(
           textStyle: SBBTextStyles.smallLight,
-          foregroundColor: baseStyle.themeValue(SBBColors.black, SBBColors.white),
+          foregroundColor: baseStyle.colorScheme.defaultTextColor,
           backgroundColor: baseStyle.themeValue(SBBColors.green, SBBColors.greenDark),
           borderColor: baseStyle.themeValue(SBBColors.green, SBBColors.greenDark),
           iconColor: SBBColors.white,
         ),
         information: SBBStatusStyle(
           textStyle: SBBTextStyles.smallLight,
-          foregroundColor: baseStyle.themeValue(SBBColors.black, SBBColors.white),
+          foregroundColor: baseStyle.colorScheme.defaultTextColor,
           backgroundColor: SBBColors.smoke,
           borderColor: SBBColors.smoke,
           iconColor: SBBColors.white,
